Factorized OVH API Calls
This was previously an interactor, we want to extract the
OVH API call code so we can use it outside of interactor logic.

Now we have ovh_request (generic class) extracted from
ovh_api_interactor (the interactor using this class, needed by
other interactors)
